# Reduce ProgramCache write lock contention