#### Problem

`ProgramCache`'s current locking behavior is needlessly hurting the unified scheduler's performance.

Unified scheduler is a new transaction scheduler for the block verification (i.e. the replaying stage). It doesn't employ batching at all as a design choice. Technically, the unified scheduler still is using `TransactionBatch`es but only with 1 transaction for each. That means it bares all the extra overheads, which has been amortized by batching.

Sidenote: I believe these overheads are all solvable (but not SOON except this one...). Also note that it's already 1.8x faster than the to-be-replaced `blockstore_processor` after this pr.

Among all those overhead sources, one of the most visible one is `ProgramCache`'s write-lock contentions. Currently, `ProgramCache` is write-locking _**3 times** unconditionally per 1-tx batch_ (for loading by `replenish_program_cache()`, for evictions by `load_and_execute_sanitized_transactions()`, for updated program commiting by `commit_transactions()`). so, it acutely hampers the unified scheduler concurrency.

#### Summary of Changes

Reduce `write-lock` of `ProgramCache` to bare-minimal. Now the usual case is 1 time for loading by `replenish_program_cache()`, while the worst case is still remaining at 3 times.

roughly ~5% consistent improvement. also note that `blockstore-processor` isn't affected while both are taking the same code-path.

#### perf improvements

##### before

```
# --block-verification-method unified-scheduler
ledger processed in 19 seconds, 182 ms
ledger processed in 19 seconds, 375 ms
ledger processed in 19 seconds, 460 ms

# --block-verification-method blockstore-processor
ledger processed in 32 seconds, 845 ms
ledger processed in 32 seconds, 786 ms
ledger processed in 32 seconds, 853 ms
```

##### after

after-the-pr's speedup-factor is now 1.8x:

```
irb(main):001:0> 32.691/17.970
=> 1.8191986644407347
```

```
# --block-verification-method unified-scheduler
ledger processed in 17 seconds, 913 ms
ledger processed in 17 seconds, 970 ms
ledger processed in 18 seconds, 20 ms

# --block-verification-method blockstore-processor
ledger processed in 32 seconds, 773 ms
ledger processed in 32 seconds, 691 ms
ledger processed in 32 seconds, 451 ms
```

context: extracted from https://github.com/anza-xyz/agave/pull/593

cc: @apfitzge