@apfitzge 2024-04-26
Changes seem fine to me, but I'm less familiar with program-cache stuff. In terms of lock contention, I just ran banking-bench as a sanity check - no programs should be compiled but if locks are grabbed less often might expect small boost in throughput.

Results seem to show around the same if not slightly better:
```
without:
{'name': 'banking_bench_total', 'median': '43963.39'}
{'name': 'banking_bench_tx_total', 'median': '45240.95'}
{'name': 'banking_bench_success_tx_total', 'median': '43159.22'}

{'name': 'banking_bench_total', 'median': '44160.32'}
{'name': 'banking_bench_tx_total', 'median': '45466.70'}
{'name': 'banking_bench_success_tx_total', 'median': '43331.09'}

{'name': 'banking_bench_total', 'median': '46828.73'}
{'name': 'banking_bench_tx_total', 'median': '48264.15'}
{'name': 'banking_bench_success_tx_total', 'median': '46307.88'}

with:
{'name': 'banking_bench_total', 'median': '41763.08'}
{'name': 'banking_bench_tx_total', 'median': '42942.16'}
{'name': 'banking_bench_success_tx_total', 'median': '40872.13'}

{'name': 'banking_bench_total', 'median': '48499.97'}
{'name': 'banking_bench_tx_total', 'median': '50037.59'}
{'name': 'banking_bench_success_tx_total', 'median': '48014.97'}

{'name': 'banking_bench_total', 'median': '46336.45'}
{'name': 'banking_bench_tx_total', 'median': '47713.88'}
{'name': 'banking_bench_success_tx_total', 'median': '45633.68'}
```